WIP show enrollments in published and unpublished sections
